refactor(unwind): Generic closer
