+++ vms/saevm/cchain/BUILD.bazel
@@ -28,7 +28,6 @@ go_library(
         "gossip.go",
         "hooks.go",
         "metrics.go",
-        "state.go",
         "vm.go",
         "warp.go",
     ],
@@ -70,6 +69,7 @@ go_library(
         "//utils/logging",
         "//utils/rpc",
         "//utils/set",
+        "//utils/unwind",
         "//vms",
         "//vms/components/avax",
         "//vms/components/gas",
@@ -120,7 +120,6 @@ go_test(
         "gossip_test.go",
         "hooks_test.go",
         "metrics_test.go",
-        "state_test.go",
         "vm_test.go",
         "warp_test.go",
     ],
